Move files to mips/normrange.
